Split autodetector tests into separate files
